requirement.txt
numpy
plotly.graph_objs
cgi